Refactor card creation into new file
